Separate `eval.c` test.
